Update and simplifiy component unit tests
